creating all price validation and adding the sku column product

+++ inc/Controller/ProductController.php
@@ -28,7 +28,7 @@ class ProductController{
         global $wpdb;
 
         $products = $wpdb->get_results(
-            $wpdb->prepare("SELECT * FROM $wpdb->prefix" . "posts  LEFT JOIN $wpdb->prefix" . "postmeta ON ID=post_id WHERE post_type IN (%s,%s) AND post_status NOT IN (%s,%s) AND meta_key = %s", 'product','product_variation','auto-draft','trash')
+            $wpdb->prepare("SELECT * FROM $wpdb->prefix" . "posts  LEFT JOIN $wpdb->prefix" . "postmeta ON ID=post_id WHERE post_type IN (%s,%s) AND post_status NOT IN (%s,%s) AND meta_key = %s", 'product','product_variation','auto-draft','trash','_sku')
         );
 
         $products = stdToArray($products);
@@ -37,7 +37,8 @@ class ProductController{
            if(! $this->isProductHasVariations($product['ID'])){ //not including products with variations
                 $image_values = wp_get_attachment_image_src( get_post_thumbnail_id($product['ID']), 'single-post-thumbnail' );
                 $product['image'] = $image_values[0];
-                $product['price'] = $product['meta_value'];
+                $product['price'] = $this->getRegularPrice($product['ID']);
+                $product['sku'] = $product['meta_value'];
                 $product['sale_price'] = $this->getSalesPrice($product['ID']);
                 array_push($product_with_its_variations,$product);
             }
@@ -51,6 +52,19 @@ class ProductController{
 
         $products = $wpdb->get_results(
             $wpdb->prepare("SELECT * FROM $wpdb->prefix" . "postmeta WHERE post_id = %d AND meta_key = %s", $id ,'_sale_price')
+        );
+        $products = stdToArray($products);
+        if(count($products)>0){
+            return $products[0]['meta_value'];
+        }
+        return 0;
+    }
+
+    function getRegularPrice($id){
+        global $wpdb;
+
+        $products = $wpdb->get_results(
+            $wpdb->prepare("SELECT * FROM $wpdb->prefix" . "postmeta WHERE post_id = %d AND meta_key = %s", $id ,'_regular_price')
         );
         $products = stdToArray($products);
         if(count($products)>0){
@@ -120,24 +134,17 @@ class ProductController{
         $price = $_POST['price'];
         $sale_price = $_POST['sale_price'];
 
-       // $price_old = get_post_meta($post_id,'_regular_price');
-
-        $result = array('failure' => 'ERROR');
-
-        update_post_meta($post_id, '_sale_price', $sale_price);
-        update_post_meta($post_id, '_regular_price', $price);
-
-        if($sale_price == 0){
-            delete_post_meta($post_id, '_sale_price');
-            update_post_meta($post_id,'_price',$price);
-
-            $result = array('success' => 'Price updated');
-        }if($sale_price < $price && $sale_price > 0){
+        if($sale_price < $price && $sale_price > 0 ){ //S<R and S>0
             update_post_meta($post_id, '_regular_price', $price);
             update_post_meta($post_id, '_price', $sale_price);
-            $result = array('success' => 'Price updated');
+            update_post_meta($post_id, '_sale_price', $sale_price);
+            $result = array('success' => 'Sale Price Updated and Regular Price Updated');
+        }else{
+            update_post_meta($post_id, '_regular_price', $price);
+            update_post_meta($post_id, '_price', $price);
+            delete_post_meta($post_id, '_sale_price');
+            $result = array('success' => 'Regular Price Updated and Sale Price Deleted');
         }
-
         echo json_encode($result);
         wp_die();
     }
